% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locker.R
\name{clear_locker}
\alias{clear_locker}
\title{Clears the locker space}
\usage{
clear_locker(where, pattern)
}
\arguments{
\item{where}{The locker directory.}

\item{pattern}{A regular expression for selecting files to clear.}
}
\description{
Clears the locker space
}
\details{
Because this actually clears files, we validate here to make sure this
isn't called by accident on the wrong directory.
}
