% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locker.R
\name{config_locker}
\alias{config_locker}
\title{Configure a locker directory}
\usage{
config_locker(where, ask = NULL, noreset = NULL)
}
\arguments{
\item{where}{The locker location.}

\item{ask}{Logical; toggles whether or not the user will be asked to confirm
prior to resetting a locker space.}

\item{noreset}{Logical; toggles whether or not a locker space is resettable.}
}
\description{
Configure a locker directory
}
