% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst.R
\name{internalize_fst}
\alias{internalize_fst}
\alias{get_fst}
\title{Get the contents of an fst file set}
\usage{
internalize_fst(path, ...)

get_fst(path, ...)
}
\arguments{
\item{path}{The directory to search.}

\item{...}{Other arguments passed to \code{\link[fst:write_fst]{fst::read_fst()}}.}
}
\description{
Get the contents of an fst file set
}
\seealso{
\code{\link[=list_fst]{list_fst()}}, \code{\link[=head_fst]{head_fst()}}
}
