% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA.R
\docType{package}
\name{paquet-package}
\alias{paquet}
\alias{paquet-package}
\title{paquet: File Streams for Big Data}
\description{
Provides tools for creating and managing file streams in support of large simulation or other outputs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://metrumresearchgroup.github.io/paquet}
  \item \url{https://github.com/metrumresearchgroup/paquet}
  \item Report bugs at \url{https://github.com/metrumresearchgroup/paquet/issues}
}

}
\author{
\strong{Maintainer}: Kyle T Baron \email{kyleb@metrumrg.com} (\href{https://orcid.org/0000-0001-7252-5656}{ORCID})

Other contributors:
\itemize{
  \item Dan Polhamus \email{danp@metrumrg.com} [contributor]
  \item Seth Green \email{sethg@metrumrg.com} [contributor]
  \item Kyle Meyer \email{kylem@metrumrg.com} [contributor]
}

}
\keyword{internal}
