% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locker.R
\name{require_ask_locker}
\alias{require_ask_locker}
\alias{no_ask_locker}
\title{Manage ask status for a locker space}
\usage{
require_ask_locker(where)

no_ask_locker(where)
}
\arguments{
\item{where}{The locker directory.}
}
\value{
Invisible \code{NULL}.
}
\description{
Use \code{require_ask_locker} to place a hidden file which will require user
confirmation every time a locker reset attempt is made. Use \code{no_ask_locker}
to remove that file so that the user will no longer be asked.
}
\details{
An error will be generated if \code{where} isn't already a valid locker space.
}
\examples{
dir <- file.path(tempdir(), "my-locker")
reset_locker(dir)
require_ask_locker(dir)
list.files(dir)

no_ask_locker(dir)
list.files(dir)

}
